+++ linode/expand.go
@@ -1,6 +1,6 @@
 package linode
 
-import "github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
 func expandStringList(list []interface{}) []string {
 	slice := make([]string, 0, len(list))