+++ linode/resource_linode_nodebalancer_config.go
@@ -7,8 +7,8 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/linode/linodego"
 )
 
@@ -137,16 +137,16 @@ func resourceLinodeNodeBalancerConfig() *schema.Resource {
 				Sensitive:   true,
 			},
 			"node_status": {
-				Type:     schema.TypeMap,
+				Type:     schema.TypeList,
 				Computed: true,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
-						"status_up": {
+						"up": {
 							Type:        schema.TypeInt,
 							Description: "The number of backends considered to be 'UP' and healthy, and that are serving requests.",
 							Computed:    true,
 						},
-						"status_down": {
+						"down": {
 							Type:        schema.TypeInt,
 							Description: "The number of backends considered to be 'DOWN' and unhealthy. These are not in rotation, and not serving requests.",
 							Computed:    true,
@@ -224,12 +224,10 @@ func resourceLinodeNodeBalancerConfigRead(d *schema.ResourceData, meta interface
 	d.Set("ssl_key", config.SSLKey)
 	d.Set("ssl_fingerprint", config.SSLFingerprint)
 	d.Set("ssl_commonname", config.SSLCommonName)
-
-	nodeStatus := map[string]interface{}{
-		"up":   fmt.Sprintf("%d", config.NodesStatus.Up),
-		"down": fmt.Sprintf("%d", config.NodesStatus.Down),
-	}
-	d.Set("node_status", nodeStatus)
+	d.Set("node_status", []map[string]interface{}{{
+		"up":   config.NodesStatus.Up,
+		"down": config.NodesStatus.Down,
+	}})
 
 	return nil
 }