+++ linode/resource_linode_volume.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/linode/linodego"
 )
 
@@ -138,15 +138,11 @@ func resourceLinodeVolumeCreate(d *schema.ResourceData, meta interface{}) error 
 	}
 
 	d.SetId(fmt.Sprintf("%d", volume.ID))
-	d.SetPartial("label")
-	d.SetPartial("region")
-	d.SetPartial("size")
 
 	if createOpts.LinodeID > 0 {
 		if _, err := client.WaitForVolumeLinodeID(context.Background(), volume.ID, linodeID, int(d.Timeout(schema.TimeoutUpdate).Seconds())); err != nil {
 			return err
 		}
-		d.SetPartial("linode_id")
 	}
 
 	if _, err = client.WaitForVolumeStatus(context.Background(), volume.ID, linodego.VolumeActive, int(d.Timeout(schema.TimeoutCreate).Seconds())); err != nil {
@@ -183,7 +179,6 @@ func resourceLinodeVolumeUpdate(d *schema.ResourceData, meta interface{}) error 
 		}
 
 		d.Set("size", size)
-		d.SetPartial("size")
 	}
 
 	updateOpts := linodego.VolumeUpdateOptions{}
@@ -208,9 +203,7 @@ func resourceLinodeVolumeUpdate(d *schema.ResourceData, meta interface{}) error 
 			return err
 		}
 		d.Set("tags", volume.Tags)
-		d.SetPartial("tags")
 		d.Set("label", volume.Label)
-		d.SetPartial("label")
 	}
 
 	var linodeID *int
@@ -255,7 +248,6 @@ func resourceLinodeVolumeUpdate(d *schema.ResourceData, meta interface{}) error 
 		}
 
 		d.Set("linode_id", linodeID)
-		d.SetPartial("linode_id")
 	}
 	d.Partial(false)